requirements.txt
torch==2.4.1
torchvision==0.19.1
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pillow>=10.0,<11.0
matplotlib>=3.5.0,<4.0
tqdm>=4.60.0,<5.0
numpy>=1.21.0
scikit-learn>=1.0.0
requests>=2.25.0